#206 Mobile Account Refactor - Got rid of externalized deposit and withdraw component - Got rid of unused code - Got rid of dropdown menu
